Update and rename 2016Exam.py to ListComprehention1.py
Define a	function shiftlist(alist, n) that	moves	the objects in a	list	alist to	the right	with	n positions.	For	example:	shiftlist([1,	2,	3,	4,	5],	1) should	return the	list	[5,	1, 2,	3,	4]. Note:	take	into	account	that the	argument n can be	larger	than	the	length of	your	list:	i.e.	shiftlist([1,	2,	3,	4,	5],	1) and	shiftlist([1,	2,	3,	4,	5],	6) should produce	the	same	result